requirements.txt
asyncio
streamlit
python-telegram-bot
